requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0a3
attrs==21.2.0
charset-normalizer==2.0.7
flake8==4.0.1
frozenlist==1.2.0
hat-aio==0.5.1
hat-asn1==0.5.1
hat-drivers==0.5.2
hat-json==0.5.2
hat-peg==0.5.1
hat-util==0.5.4
idna==3.3
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.1.2
mccabe==0.6.1
multidict==5.2.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyrsistent==0.18.0
pyserial==3.5
PyYAML==6.0
typing-extensions==3.10.0.2
yarl==1.7.0
